+++ lib/src/alert_dialog/show_ok_alert_dialog.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 /// This is convenient wrapper of [showAlertDialog].
 /// [barrierDismissible] (default: true) only works for material style,
 /// and if it is set to false, pressing OK button is only way to close alert.
+/// [actionsOverflowDirection] works only for Material style currently.
 Future<OkCancelResult> showOkAlertDialog({
   @required BuildContext context,
   String title,
@@ -15,6 +16,7 @@ Future<OkCancelResult> showOkAlertDialog({
   AdaptiveStyle alertStyle = AdaptiveStyle.adaptive,
   bool useActionSheetForCupertino = false,
   bool useRootNavigator = true,
+  VerticalDirection actionsOverflowDirection = VerticalDirection.up,
 }) async {
   final result = await showAlertDialog<OkCancelResult>(
     context: context,
@@ -24,6 +26,7 @@ Future<OkCancelResult> showOkAlertDialog({
     style: alertStyle,
     useActionSheetForCupertino: useActionSheetForCupertino,
     useRootNavigator: useRootNavigator,
+    actionsOverflowDirection: actionsOverflowDirection,
     actions: [
       AlertDialogAction(
         label: okLabel ?? MaterialLocalizations.of(context).okButtonLabel,